Allow Multiple Concurrent Mirrors
This is a change I've been meaning to make for a while now.

Now you can build a few mirrors at the same time.
In general, though, I've used this opportunity to tighten everything.

There's more error checking than there was, it refuses certain things.

I've tried to make it "backward compatible" in that if you only ever
have one at a time, it will work as it previously did.

It does that by picking the only one if there's only one.
It also tries to clean up more than it used to.

Specifically, if you clean up the last mirror, it should be as though
you never had any.