Added Creation Callbacks
So, if certain scripts exist, they will be run with the mirror is built.

This is something I've wanted for quite a while.
There are a series of things I always did as soon as I built a mirror so
I could tell it apart and use it and stuff.

Now this is all in scripts, so it'll just be done for me.
Happy.